refactor: merge getPageAtIndex back into waitForPage
